% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_evinb.R
\name{bootrun_evinb}
\alias{bootrun_evinb}
\title{Running an extreme value inflated negative binomial model with bootstrapping}
\usage{
bootrun_evinb(evinb, block = NULL, timing = T)
}
\arguments{
\item{formula_nb}{Formula for the negative binomial (count) component of the model}

\item{formula_zi}{Formula for the zero-inflation component of the model}

\item{formula_evi}{Formula for the extreme-value inflation component of the model}

\item{formula_pareto}{Formula for the pareto (extreme value) component of the model}

\item{data}{Data to run the model on}
}
\value{
An object of class 'evi' containing XX
}
\description{
Running an extreme value inflated negative binomial model with bootstrapping
}
\examples{
run_evinb(y~x1+x2,y~x1+x3,y~x1,y~x4, data=data_test)
}
