% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_extractor.R
\name{counts_from_evzinb}
\alias{counts_from_evzinb}
\title{Extracting fitted count values of the NB component of an evzinb object}
\usage{
counts_from_evzinb(evzinb, newdata = NULL, return_data = F)
}
\arguments{
\item{evzinb}{An evzinb object for which to produce counts}

\item{newdata}{Optional new data (tibble) to produce predicted quantiles for}

\item{return_data}{Logical: Should the data be returned in the object}
}
\value{
A tibble with the predicted nb counts. If return_data=T this is appended to the data or newdata
}
\description{
Extracting fitted count values of the NB component of an evzinb object
}
