% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVI-package.R
\docType{package}
\name{EVI-package}
\alias{EVI}
\alias{EVI-package}
\title{EVI: Epidemic Volatility Index as an Early-Warning Tool}
\description{
This is an R package implementing the epidemic volatility index (EVI), as 
    discussed by Kostoulas et. al. (in press). The package is currently in
    development.
}
\author{
\strong{Maintainer}: Polychronis Kostoulas \email{pkost@uth.gr}

Authors:
\itemize{
  \item Eletherios Meletis \email{emeletis@outlook.com}
  \item Konstantinos Pateras \email{kostas.pateras@gmail.com}
  \item Paolo Eusebi \email{paoloeusebi@gmail.com}
  \item Matt Denwood \email{md@sund.ku.dk}
}

}
\keyword{internal}
