% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVI.compare.R
\name{evi.compare}
\alias{evi.compare}
\title{Combining multiple epidemic volatility indices into a multi-early warning index plot}
\usage{
evi.compare(
  EVI1_output,
  EVI2_output,
  ln = T,
  type = "p",
  size.evi = 1,
  lg = T,
  gray = F,
  EVI1.lab = "EVI1",
  EVI2.lab = "EVI2",
  EVI3.lab = "EVI+",
  EVI.country = NULL
)
}
\arguments{
\item{EVI1_output}{e.g. output of the \code{\link[EVI:deviant]{deviant()}} function}

\item{EVI2_output}{e.g. output of the \code{\link[EVI:deviant]{deviant_cEVI()}} function}

\item{ln}{TRUE or FALSE; If TRUE (the default) the output of the graph will be presented on the logarithmic scale. IF FALSE the output data will be presented on the original scale.}

\item{type}{By default, points are plotted on EVI graphs (type="p"). In cases where, changes are very sudden or data sparsely available, type="l" introduces lines on top of points for the "EVI" type of graph.}

\item{size.evi}{Argument that controls the size of the evi lines (numeric). Default size.evi=1.}

\item{lg}{If TRUE then add the names of the 2 indexes in EVI1.lab and EVI2.lab and their combination in EVI3.lab. If FALSE, then the name of the country will be printed.}

\item{gray}{If TRUE then grayscale alternatives of the EVI.graphs will be ploted.}

\item{EVI1.lab}{Add the name of the first index as character.}

\item{EVI2.lab}{Add the name of the second index as character.}

\item{EVI3.lab}{Add the name of the combination of the first and second index as character.}
}
\value{
\itemize{
\item{evi.compare: A graph with two EVI indexes shown simulteounsly.}}
}
\description{
This function returns an evi.graph containing two indices.
\enumerate{
\item If index number 1 results in an early warning,then a yellow point will be shown.
\item If index number 2 results in an early warning, then a orange point will be shown.
\item If both indexes result in early warnings, then a red point will be shown.
}
}
\details{
For each time point the stored variables are:
}
\examples{

# Run basic functions to acquire early warnings results
EVI_output=deviant(new_cases=Italy$Cases, cum=FALSE)
cEVI_output=deviant_cEVI(new_cases=Italy$Cases, cum=FALSE)

# Plot the EVI combination graph
evi.compare(EVI1_output=EVI_output, EVI2_output=cEVI_output, ln=T)
evi.compare(EVI1_output=EVI_output, EVI2_output=cEVI_output, ln=T, type="l") # For the line EVI plot
}
\references{
Pateras K, Meletis E, Denwood M, Paolo E, Kostoulas P, The convergence epidemic volatility index an early warning tool for identifying waves in an epidemic, 2022
}
