% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi.graphs.comb.R
\name{evi.graphs.comb}
\alias{evi.graphs.comb}
\title{Combining multiple epidemic volatility indices into a multi-early warning index plot}
\usage{
evi.graphs.comb(
  EVI1_output,
  EVI2_output,
  ln = T,
  type = "p",
  size.evi = 1,
  EVI1.lab = "EVI1",
  EVI2.lab = "EVI2",
  EVI3.lab = "EVI+",
  EVI.country = NULL
)
}
\arguments{
\item{EVI1_output}{e.g. output of the \code{\link[EVI:deviant]{deviant()}} function}

\item{EVI2_output}{e.g. output of the \code{\link[EVI:deviant]{deviant_cEVI()}} function}

\item{ln}{TRUE or FALSE; If TRUE (the default) the output of the graph will be presented on the logarithmic scale. IF FALSE the output data will be presented on the original scale.}

\item{type}{By default, points are plotted on EVI graphs. In cases where, changes are very sudden or data sparsely available, type="l" introduces lines on top of points for the "EVI" type of graph.}
}
\value{
\itemize{
\item{comb.evi.graph: A graph with two EVI indexes shown simulteounsly.}}
}
\description{
This function returns an evi.graph containing two indices.
\enumerate{
\item If index number 1 results in an early warning,then a yellow point will be shown.
\item If index number 2 results in an early warning, then a orange point will be shown.
\item If both indexes result in early warnings, then a red point will be shown.
}
}
\details{
For each time point the stored variables are:
}
\examples{

# Run basic functions to acquire early warnings results
EVI_output=deviant(new_cases=Italy$Cases, cum=FALSE)
cEVI_output=deviant_cEVI(new_cases=Italy$Cases, cum=FALSE)

# Plot the EVI combination graph
evi.graphs.comb(EVI1_output=EVI_output, EVI2_output=cEVI_output, ln=T)
evi.graphs.comb(EVI1_output=EVI_output, EVI2_output=cEVI_output, ln=T, type="l") # For the line EVI plot
}
\references{
Pateras K, Meletis E, Denwood M, Paolo E, Kostoulas P, The convergence epidemic volatility index an early warning tool for identifying waves in an epidemic, 2022
}
