% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evifcut_cEVI.R
\name{evifcut_cEVI}
\alias{evifcut_cEVI}
\title{Function that calculates the sensitivity and the specificity for each cut-off value and rolling window size based on cEVI}
\usage{
evifcut_cEVI(cevi, cases, cut, r)
}
\arguments{
\item{cevi}{numeric vector - object that corresponds to either a 0 for no an early warning or an 1 for an early warning based on cEVI.}

\item{cases}{numeric vector with the number of new cases per unit of time (i.e., daily).}

\item{cut}{threshold value (0 <= c <= 0.5) for issuing an early warning. If evi >= c, an early warning is issued and otherwise is not.}

\item{r}{Definition for the minimum difference in the mean number of cases, one week before and after each time point that, if present, should be detected. This is the case definition and the default is 0.2 (with 0 <= r <= 1). A value of r=0.2 means that we have a case when the mean number of the newly observed cases in the next 7 days is at least 20\% higher than the mean number of the newly observed cases in the past 7 days.}
}
\description{
Function that calculates the sensitivity and the specificity for each cut-off value and rolling window size.
}
\examples{
data("Italy")
cases = mova(cases=Italy$Cases)
evifcut_cEVI(cevi=cevi, cases=cases, cut=0.01, r=0.2)
}
