% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example}
\alias{example}
\title{Example function and help file}
\usage{
example(argument = TRUE)
}
\arguments{
\item{argument}{explanation of argument}
}
\description{
Example function and help file
}
\examples{
example(TRUE)
example(FALSE)

}
