% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_extractor.R
\name{fitted_alpha_from_evzinb}
\alias{fitted_alpha_from_evzinb}
\title{Extracting fitted alpha values of the pareto component of an evzinb object}
\usage{
fitted_alpha_from_evzinb(evzinb, newdata = NULL, return_data = F)
}
\arguments{
\item{evzinb}{An evzinb object for which to produce quantiles}

\item{newdata}{Optional new data (tibble) to produce predicted quantiles for}

\item{return_data}{Logical: Should the data be returned in the object}
}
\value{
A tibble with the predicted pareto alpha values. If return_data=T this is appended to the data or newdata
}
\description{
Extracting fitted alpha values of the pareto component of an evzinb object
}
