% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_evzinb.R
\name{glance.evzinb}
\alias{glance.evzinb}
\title{EVZINB tidy function}
\usage{
\method{glance}{evzinb}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}

\item{coef}{}

\item{p_value}{}

\item{confint}{}

\item{component}{}
}
\value{
An EVZINB glance function
}
\description{
EVZINB tidy function
}
\examples{
tidy(evzinb)
}
