% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic.R
\name{indic}
\alias{indic}
\title{Description to be added}
\usage{
indic(evi, cut, cases)
}
\arguments{
\item{evi}{explanation of argument}

\item{cut}{explanation}

\item{cases}{explanation}
}
\description{
Description to be added
}
\examples{
indic(evi, cut, cases)

}
