% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_cEVI.R
\name{indic_cEVI}
\alias{indic_cEVI}
\title{Function that produces the early warning based on cEVI}
\usage{
indic_cEVI(cevi, cut, cases)
}
\arguments{
\item{cevi}{numeric vector - object that corresponds to either a 0 for no an early warning or an 1 for an early warning based on cEVI.}

\item{cut}{threshold value (0 <= c <= 0.5) for issuing an early warning. If evi >= c an early warning is issued and otherwise is not.}

\item{cases}{numeric vector with the number of new cases per unit of time (i.e., daily).}
}
\description{
Index = 1 when an early warning is issued and Index=0 when an early warning is not issued
}
\examples{
data("Italy")
cases = mova(cases=Italy$Cases, r_a=7)
ind=indic_cEVI(cevi=cev, cut=0.01, cases=cases)

}
