% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Italy.R
\docType{data}
\name{Italy}
\alias{Italy}
\title{Cases of the first 149 days of the Covid-19 pandemic in Italy.}
\format{
A data frame with 149 rows and 3 variables:
\describe{
\item{Date}{ID of date encoded with origin="01-01-1970"}
\item{Cases}{Number of newly observed cases per day}
\item{Cum_Cases}{Cumulative number of newly observed cases per day}
}
}
\source{
\url{https://github.com/CSSEGISandData/COVID-19}
}
\usage{
Italy
}
\description{
A Dataframe containing the number of cases for the first 149 days of the Covid-19 pandemic
in Italy
}
\keyword{datasets}
