% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medvol.R
\name{medvol}
\alias{medvol}
\title{Function that calculates the standard deviation}
\usage{
medvol(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Calculates and returns the standard deviation of a vector
}
\examples{
x=c(2,2,1,4,9,10,23,10,9,10,14,12,10)
medvol(x)
5.922577

}
