% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mova.R
\name{mova}
\alias{mova}
\title{Example function and help file}
\usage{
mova(cases, int = 7)
}
\arguments{
\item{cases}{time series data}

\item{int}{rolling window}
}
\description{
Example function and help file
}
\examples{
cases=mova(c(2,2,1,4,9,10,23,10,9,10,14,12,10),7)

}
