% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_extractor.R
\name{prob_from_evzinb}
\alias{prob_from_evzinb}
\title{Extracting state probabilities from an evzinb object}
\usage{
prob_from_evzinb(evzinb, newdata = NULL, return_data = F)
}
\arguments{
\item{evzinb}{An evzinb object for which to produce probabilities}

\item{newdata}{Optional new data (tibble) to produce predicted quantiles for}

\item{return_data}{Logical: Should the data be returned in the object}
}
\value{
A tibble with the predicted state probabilities. If return_data=T this is appended to the data or newdata
}
\description{
Extracting state probabilities from an evzinb object
}
