% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_extractor.R
\name{quantiles_from_evzinb}
\alias{quantiles_from_evzinb}
\title{Extracting full mixture quantiles from an evzinb object}
\usage{
quantiles_from_evzinb(
  evzinb,
  quantile,
  newdata = NULL,
  return_data = F,
  multicore_ok = T,
  ncores = NULL
)
}
\arguments{
\item{evzinb}{An evzinb object for which to produce quantiles}

\item{quantile}{The quantile for which to produce predictions}

\item{newdata}{Optional new data (tibble) to produce predicted quantiles for}

\item{return_data}{Logical: Should the data be returned in the object}

\item{multicore_ok}{Logical: Should parallel processing be used to obtain results}

\item{ncores}{Number of cores if multicore is used}
}
\value{
A vector of predicted quantiles, or if return_data=T, a tibble with the predicted quantile attached last
}
\description{
Extracting full mixture quantiles from an evzinb object
}
\examples{
quantiles_from_evzinb(evzinb,0.95)
}
