% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_evzinb.R
\name{run_evzinb}
\alias{run_evzinb}
\title{Running an extreme value and zero inflated negative binomial model}
\usage{
run_evzinb(
  formula_nb,
  formula_zi,
  formula_evi,
  formula_pareto,
  data,
  max.diff.par = 0.001,
  max.no.em.steps = 200,
  max.no.em.steps.warmup = 5,
  c.lim = c(50, 1000),
  max.upd.par.zc.multinomial = 0.5,
  max.upd.par.pl.multinomial = 0.5,
  max.upd.par.nb = 0.5,
  max.upd.par.pl = 0.5,
  no.m.bfgs.steps.multinomial = 3,
  no.m.bfgs.steps.nb = 3,
  no.m.bfgs.steps.pl = 3,
  pdf.pl.type = "approx",
  eta.int = c(-1, 1),
  init.Beta.multinom.ZC = NULL,
  init.Beta.multinom.PL = NULL,
  init.Beta.NB = NULL,
  init.Beta.PL = NULL,
  init.Alpha.NB = 0.01,
  init.C = 200
)
}
\arguments{
\item{formula_nb}{Formula for the negative binomial (count) component of the model}

\item{formula_zi}{Formula for the zero-inflation component of the model}

\item{formula_evi}{Formula for the extreme-value inflation component of the model}

\item{formula_pareto}{Formula for the pareto (extreme value) component of the model}

\item{data}{Data to run the model on}
}
\value{
An object of class 'evi' containing XX
}
\description{
Running an extreme value and zero inflated negative binomial model
}
\examples{
run_evzinb(y~x1+x2,y~x1+x3,y~x1,y~x4, data=data_test)
}
