% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfun.R
\name{testfun}
\alias{testfun}
\title{A function for testing our package}
\usage{
testfun(x, y, z)
}
\arguments{
\item{x}{a scalar to add}

\item{y}{another scalar to add}

\item{z}{a third scalar to add}
}
\value{
a scalar, the sum of x, y and z
}
\description{
A function for testing our package
}
\examples{
testfun(1,2,3)
}
