% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_evzinb.R
\name{tidy.evzinb}
\alias{tidy.evzinb}
\title{EVZINB tidy function}
\usage{
\method{tidy}{evzinb}(
  x,
  component = c("zi", "evi", "count", "pareto", "all"),
  coef = c("original", "bootstrapped_mean", "bootstrapped_median"),
  standard_error = TRUE,
  p_value = c("bootstrapped", "approx", "none"),
  confint = c("none", "bootstrapped", "approx"),
  conf_level = 0.95,
  approx_t_value = TRUE,
  symmetric_bootstrap_p = TRUE,
  ...
)
}
\arguments{
\item{x}{}

\item{component}{}

\item{coef}{}

\item{p_value}{}

\item{confint}{}

\item{...}{}
}
\value{
An EVZINB tidy function
}
\description{
EVZINB tidy function
}
\examples{
tidy(evzinb)
}
