% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviant.R
\name{deviant}
\alias{deviant}
\title{The main function of the EVI package that utilizes all secondary functions and the evi rolling function to give early warnings.}
\usage{
deviant(
  new_cases,
  ratio = 1/1.2,
  lag_1 = 7,
  lag_max = 30,
  start_cases = 14,
  c_1 = 7,
  w_s = 7,
  c_s = seq(0.01, 0.5, 0.01)
)
}
\arguments{
\item{new_cases}{explanation}

\item{ratio}{explanation}

\item{lag_1}{initial lag}

\item{lag_max}{maximum lag}

\item{start_cases}{explanation}

\item{c_1}{initial cut of point}

\item{w_s}{desired detection difference}

\item{c_s}{cut off point range}
}
\description{
The main function of the EVI package that utilizes all secondary functions and the evi rolling function to give early warnings.
}
\examples{
deviant(new_cases, ratio, lag_max, start_cases, lag_1, c_1, w_s, c_s)

}
