% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviant.R
\name{deviant}
\alias{deviant}
\title{The main function of the EVI package that utilizes all secondary functions and the evi rolling function to give early warnings.}
\usage{
deviant(new_cases, ratio = 1/1.2, lag_max = 30, start_cases = 14)
}
\arguments{
\item{new_cases}{numeric vector calculated from the mova function?}

\item{ratio}{threshold value for case definition - ratio=1/1.2{default} as defined above}

\item{lag_max}{maximum value for the rolling window size lag_max=30{default}}

\item{start_cases}{interget - first observations "burn-in" of the time series}

\item{lag_1}{(Cannot be changed) initial value for the rolling windows size, lag_1=7,}

\item{c_1}{(Cannot be changed)initital value for the threshold value of model prediction, c_1=0.1}

\item{w_s}{(Cannot be changed) desired detection difference, w_s=7}

\item{c_s}{(Cannot be changed) cut off point range, c_s=seq(0.01,0.5, 0.01)}
}
\description{
From the indic and status functions a 2x2 table can be constructed -
indic - represents the model prediction at the specified cut-off value
status - represents the cases (true cases)
Case definition - Status function
Indic function output |               1            |       0
at cut-off value c    |  1         Se*prev         |     (1-Sp) * (1-prev)
Model prediction      |  0         (1-Se)*prev     |    Sp * (1-prev)
}
\details{
Optimization function that finds that optimal cut-off value that maximizes the Youden index and based on that cut-off value calculates PPN, NPV etc
}
\examples{
load(Itdat)
ratio = 1/1.2
lag_max=30
start_cases=14
deviant(Itdat, ratio, lag_max, start_cases) - this step will take some time

}
