% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi-package.R
\docType{package}
\name{evi-package}
\alias{evi}
\alias{evi-package}
\title{evi: Inference with extreme value inflated count data}
\description{
This package allows users to model and draw inferences from extreme value inflated count data, and to evaluate these models and compare to non extreme-value inflated counterparts. The package is built to be compatible with standard presentation tools such as broom, tidy, and modelsummary.
}
\author{
\strong{Maintainer}: David Randahl \email{david.randahl@pcr.uu.se}

Authors:
\itemize{
  \item Johan Vegelius
}

}
\keyword{internal}
