% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVI-package.R
\docType{package}
\name{EVI-package}
\alias{EVI}
\alias{EVI-package}
\title{EVI: Epidemic Volatility Index as an Early-Warning Tool}
\description{
Implementation of the epidemic volatility index (EVI), as discussed by Kostoulas et. al. (2021). EVI is a new, conceptually simple, early warning tool for oncoming epidemic waves. EVI is based on the volatility of newly reported cases per unit of time, ideally per day, and issues an early warning when the volatility change rate exceeds a threshold.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.nature.com/articles/s41598-021-02622-3}
}

}
\author{
\strong{Maintainer}: Eletherios Meletis \email{emeletis@outlook.com}

Authors:
\itemize{
  \item Konstantinos Pateras \email{kostas.pateras@gmail.com}
  \item Paolo Eusebi \email{paoloeusebi@gmail.com}
  \item Matt Denwood \email{md@sund.ku.dk}
  \item Polychronis Kostoulas \email{pkost@uth.gr}
}

}
\keyword{internal}
