% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVI.graphs.R
\name{evi.graphs}
\alias{evi.graphs}
\title{Descriptive figures demonstrating the predictive capabilities of EVI.}
\usage{
evi.graphs(EVI_output, graph = c("cases"), ln = T)
}
\arguments{
\item{graph}{Type of graph to be plotted. Default is cases, but "pv_plus" or "pv_minus" can also be chosen.}

\item{ln}{If the data would be transformed through the natural logarithm before plotting.}
}
\description{
Descriptive figures demonstrating the predictive capabilities of EVI.
}
\examples{
evi.plot(graph=c("cases",ln=T)) #

}
