% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi.R
\name{evi}
\alias{evi}
\title{Calulation of the Epidemic Volatility Index}
\usage{
evi(rollsd)
}
\arguments{
\item{rollsd}{numeric vector - returned and stored as roll from the rollsd() function}
}
\description{
Calculates the relative change in the standard deviation between two consecutive rolling windows
}
\examples{
data("Italy")
cases = mova(cases=Italy$Cases)
roll = rollsd(cases=cases)
ev = evi(rollsd=roll)

}
