% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi.R
\name{evi}
\alias{evi}
\title{Epidemic Volatility Index calculation}
\usage{
evi(rollsd)
}
\arguments{
\item{rollsd}{numeric vector - returned and stored as roll from the rollsd() function}
}
\description{
Calculates the relative change of standard deviation between two consecutive rolling windows
}
\examples{
data(Italy)
cases = mova(Italy$ncases)
roll = rollsd(cases)
ev = evi(roll)

}
