% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evifcut.R
\name{evifcut}
\alias{evifcut}
\title{Function that calculates the sensitivity and the specificity for each cut-off value and rolling window size}
\usage{
evifcut(evi, cases, cut, r)
}
\arguments{
\item{evi}{numeric vector - object (obtained from the evi function and stored as ev) that corresponds to the relative change in the standard deviation}

\item{cases}{numeric vector with a number of cases per unit of time (i.e., daily)}

\item{cut}{threshold value (0<=c<=0.5) for issuing an early warning. If evi >= c an early warning is issued and otherwise is not.}

\item{r}{0<=r<=1, r=0.2 {default}) Definition for the minimum increase in the mean number of cases, one week before and after each time point that if present should be detected (i.e., defines a case). The default is 0.2. This means that we have a case if the mean number of the newly observed cases in the next 7 days is at least 20\% higher than the mean number of the newly observed cases in the past 7 days}
}
\description{
Function that calculates the sensitivity and the specificity for each cut-off value and rolling window size
}
\examples{
#' data("Italy")
cases = mova(cases=Italy$Cases)
roll = rollsd(cases=cases)
ev = evi(rollsd=roll)
evifcut(evi=ev, cases=cases, cut=0.01, r=0.2)
}
