% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evifcut.R
\name{evifcut}
\alias{evifcut}
\title{Description to be added (#test which cut-off c & lag)}
\usage{
evifcut(evi, cases, rate, w_s, ratio)
}
\arguments{
\item{evi}{explanation of argument}

\item{cases}{explanation}

\item{rate}{explanation}

\item{w_s}{explanation}

\item{ratio}{explanation}
}
\description{
Description to be added (#test which cut-off c & lag)
}
\examples{
evifcut(evi, cases, rate, w_s, ratio)

}
