% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evifcut.R
\name{evifcut}
\alias{evifcut}
\title{Function that calculates the sensitivity and specificity in a range of cut-off values so that to find the cut-off value that maximizes the Youden Index (Se + Sp -1)}
\usage{
evifcut(evi, cases, rate, w_s = 7, ratio = 1/1.2)
}
\arguments{
\item{evi}{numeric vector as returned from the evi function}

\item{cases}{numeric vector - number of new cases per day/any other time interval}

\item{rate}{threshold value for model prediction - test outcome # defined as c in the indic function}

\item{w_s}{time interval - validation time w_s=7{default}}

\item{ratio}{threshold value for case definition ratio=1/1.2{dedault}}
}
\description{
Function that calculates the sensitivity and specificity in a range of cut-off values so that to find the cut-off value that maximizes the Youden Index (Se + Sp -1)
}
\examples{
evi: numeric vector as returned from the evi function
cases <- rbinom(100,10,0.5)
rate = 0.01
w_s = 7
ratio = 1/1.2
evifcut(evi, cases, rate, w_s, ratio)
$sens
0.1333333
$spec
0.8611111
$testsin
0.1290323
$prev
0.1612903

}
