% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_evzinb.R
\name{glance.evzinb}
\alias{glance.evzinb}
\alias{evzinb-glance}
\alias{glance,}
\alias{evzinb-method,}
\alias{evinb-glance}
\alias{evinb-method}
\alias{glance.evinb}
\title{EVZINB and EVINB glance functions}
\usage{
\method{glance}{evzinb}(x, ...)

\method{glance}{evinb}(x, ...)
}
\arguments{
\item{x}{An EVZINB or EVINB object}

\item{...}{Further arguments to be passed to glance()}
}
\value{
An EVZINB glance function
}
\description{
EVZINB and EVINB glance functions
}
\examples{
tidy(evzinb)
}
\seealso{
\code{\link[generics]{glance}}
}
