% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic.R
\name{indic}
\alias{indic}
\title{Indicator function
\subsection{The 0 and 1 outcomes from the status and indic function are used to create the 2x2 table}{

Returns the EVI index/model prediction - 1/0
}}
\usage{
indic(evi, cut, cases)
}
\arguments{
\item{evi}{numeric vector - object (obtained from the evi function and stored as ev)}

\item{cut}{threshold value 0<=c<=1 - expetation in the future number of cases}

\item{cases}{moving average for the time series epidemic data - obtained and stored as cases from the mova function}
}
\description{
Indicator function
\subsection{The 0 and 1 outcomes from the status and indic function are used to create the 2x2 table}{

Returns the EVI index/model prediction - 1/0
}
}
\examples{
data(Italy)
cases = mova(Italy$ncases)
roll = rollsd(cases)
ev = evi(roll)
ind=indic(ev, 0.01, cases)

}
