% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mova.R
\name{mova}
\alias{mova}
\title{Function that calculates the moving average of a series of data}
\usage{
mova(cases, r_a = 7)
}
\arguments{
\item{cases}{time series data}

\item{r_a}{rolling window size/time interval on which the moving average will be calculated - number of consecutive observations per rolling window. Usually the 7-day moving average rather than the actually observed cases are analyzed}
}
\description{
Example function and help file for mova.r
Moving Average
Moving Average
Calculate the moving average for the time series epidemic data
}
\examples{
data(Italy)
7-day moving average
mova(Italy$ncases, r_a=7)
14-day moving average
mova(Italy$ncases, r_a=14)
}
