% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mova.R
\name{mova}
\alias{mova}
\title{Example function and help file for mova.r
Moving Average
Calculate the moving average for the time series epidemic data}
\usage{
mova(cases, int = 7)
}
\arguments{
\item{cases}{time series data}

\item{int}{rolling window/time interval on which the moving average will be calculated}
}
\description{
Example function and help file for mova.r
Moving Average
Calculate the moving average for the time series epidemic data
}
\examples{
cases=c(2,2,1,4,9,10,23,10,9,10,14,12,10)
mova=mova(cases, 7)
2.000000  2.000000  1.666667  2.250000
3.600000  4.666667  7.285714  8.428571
9.428571 10.714286 12.142857 12.571429
12.571429
}
