% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollsd.R
\name{rollsd}
\alias{rollsd}
\title{Rolling standard deviation}
\usage{
rollsd(cases, lag_t = 7)
}
\arguments{
\item{cases}{numeric vector - number of cases per day (or any other time interval)}

\item{lag_t}{integer - number of consecutive rolling windows (minimum/default value = 7, maximum recommended value = 30)}
}
\description{
-A function for computing the rolling standard deviation (between two sd’s) for the time series epidemic data
}
\examples{
data(Italy)
cases = mova(Italy$ncases, r_a=7)
roll = rollsd(cases,lag_t)


}
