% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollsd.R
\name{rollsd}
\alias{rollsd}
\title{Description to be added (rolling sd)}
\usage{
rollsd(cases, lag_t)
}
\arguments{
\item{cases}{explanation of argument}

\item{lag_t}{explanation of argument}
}
\description{
Description to be added (rolling sd)
}
\examples{
cases=c(2,2,1,4,9,10,23,10,9,10,14,12,10)
rollsd(cases,lag_t)

}
