% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Case definition
Case definition
\subsection{The 0 and 1 outcomes from the status and indic function are used to create the 2x2 table}{
}}
\usage{
status(cases, w_s = 7, r = 0.2)
}
\arguments{
\item{cases}{numeric vector - moving average for the time series epidemic data - obtained and stored as cases from the mova function}

\item{w_s}{(Cannot be changed) time interval - validation time w_s=7{default}}

\item{r}{Threshold value (0<=r<=1, r=0.2{default}) for the minimum increase in the mean number of cases between two consecutive weeks that if present defines a case}
}
\description{
Case definition
Case definition
\subsection{The 0 and 1 outcomes from the status and indic function are used to create the 2x2 table}{
}
}
\examples{
data(Italy)
cases = mova(Italy$ncases)
status = status(cases, 7, 0.2)
}
