% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Description to be added}
\usage{
status(cases, w_s, ratio)
}
\arguments{
\item{cases}{explanation of argument}

\item{w_s}{explanation of argument}

\item{ratio}{explanation of argument}
}
\description{
Description to be added
}
\examples{
status(cases, w_s, ratio)

}
