% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_evzinb.R
\name{summary.evinb}
\alias{summary.evinb}
\title{EVINB summary function}
\usage{
\method{summary}{evinb}(object, boot_mean = F, boot_se = F, p_approx = F, ...)
}
\arguments{
\item{object}{an EVZINB object with bootstraps}

\item{boot_mean}{Logical: Should the bootstrapped mean of the estimates be computed?}

\item{boot_se}{Logical: Should the bootstrapped standard errors of the estimates be computed?}

\item{p_approx}{Logical: Should the normal approximated p-value be computed? Not recommended unless the number of bootstraps is too low for bootstrapped p-values}
}
\value{
An EVZINB summary object
}
\description{
EVINB summary function
}
\examples{
summary(evinb)
}
